* upstream/master: (16702 commits)
  Resolve "Modify the empty state of changes in MR"
  Remove unnecessary sidebar element realignment
  Add break word wrapping to well-list
  Update CHANGELOG.md for 10.3.3
  Fix user membership destroy relation
  Use heredoc so it's more clear
  Put all menus under menu module
  Introduce common project settings and just put
  Introduce Factory::Resource::DeployKey
  Introduce expand_deploy_keys
  Introduce Menu::Side
  Add documents for GitLab utilities
  Clears visual token on second backspace
  Update prometheus gem to version that adds inf+ bucket in accordance with Prometheus docs.
  Add breadcrumbs to User Settings sub-views
  Rename asset sync related AWS variables
  Allow logged in user to change his password
  Fix 404 error after a user edits an issue description and solves the reCAPTCHA
  Fix links to old commits in merge requests
  Typos correction on the advise of @smcgivern
  ...

+++ app/services/milestones/destroy_service.rb
@@ -1,15 +1,17 @@
 module Milestones
   class DestroyService < Milestones::BaseService
     def execute(milestone)
+      return unless milestone.project_milestone?
+
       Milestone.transaction do
         update_params = { milestone: nil }
 
         milestone.issues.each do |issue|
-          Issues::UpdateService.new(project, current_user, update_params).execute(issue)
+          Issues::UpdateService.new(parent, current_user, update_params).execute(issue)
         end
 
         milestone.merge_requests.each do |merge_request|
-          MergeRequests::UpdateService.new(project, current_user, update_params).execute(merge_request)
+          MergeRequests::UpdateService.new(parent, current_user, update_params).execute(merge_request)
         end
 
         event_service.destroy_milestone(milestone, current_user)
